MLIR, LLVM: Add an IR utility to perform proper slicing
This commit introduces a backwards slicing utility that is both
`RegionBranchOpInterface` and `BranchOpInterface` aware. This is a first
step in replacing the currently broken slicing utilities that cannot be
used for structured control flow.

Note that the utility is currently placed in the IR library, because
LLVM's inlining uses it and is part of the LLVMIR library. There was
originally a substantial push back on moving the inliner interface out
of said library.